[Framework] Combine ScriptGroup and ScriptChain classes and dialogs.
Dialogs can no longer be edited unfortunately, maybe that can be worked
around somehow.
