+++ Extensions/AudioScripts/Mpdn.ScriptGroup.cs
@@ -15,151 +15,14 @@
 // License along with this library.
 
 using System;
-using System.Collections.Generic;
 using System.Linq;
-using Mpdn.Extensions.Framework;
-using Mpdn.Extensions.Framework.Chain;
-using Mpdn.Extensions.PlayerExtensions;
-using Mpdn.AudioScript;
 using Mpdn.Extensions.Framework.AudioChain;
-using YAXLib;
 
 namespace Mpdn.Extensions.AudioScripts
 {
     namespace Mpdn.ScriptGroup
     {
-        public class ScriptGroup : PresetCollection<AudioOutput, IAudioScript>
-        {
-            #region Settings
-
-            public int SelectedIndex { get; set; }
-
-            public string Hotkey
-            {
-                get { return m_Hotkey; }
-                set
-                {
-                    m_Hotkey = value ?? "";
-                    UpdateHotkey();
-                }
-            }
-
-            [YAXDontSerialize]
-            public Preset<AudioOutput, IAudioScript> SelectedOption
-            {
-                get { return Options != null ? Options.ElementAtOrDefault(SelectedIndex) : null; }
-            }
-
-            [YAXDontSerialize]
-            public string ScriptName
-            {
-                get
-                {
-                    var preset = SelectedOption;
-                    return preset == null ? string.Empty : preset.Name; 
-                }
-                set
-                {
-                    var index = Options.FindIndex(p => p.Name == value);
-                    if (index < 0)
-                    {
-                        throw new KeyNotFoundException(string.Format("ScriptName '{0}' is not found", value));
-                    }
-                    SelectedIndex = index;
-                }
-            }
-
-            #endregion
-
-            private Preset<AudioOutput, IAudioScript> m_CurrentOption;
-
-            public ScriptGroup()
-            {
-                SelectedIndex = 0;
-                m_HotkeyGuid = Guid.NewGuid();
-            }
-
-            public int GetPresetIndex(Guid guid)
-            {
-                return Options.FindIndex(o => o.Guid == guid);
-            }
-
-            public override AudioOutput Process(AudioOutput input)
-            {
-                RefreshOption();
-                return SelectedOption != null ? input + SelectedOption : input;
-            }
-
-            public override void Initialize()
-            {
-                RegisterHotkey();
-                base.Initialize();
-            }
-
-            public override void Reset()
-            {
-                DeregisterHotkey();
-                base.Reset();
-            }
-
-            private void RefreshOption()
-            {
-                if (m_CurrentOption == SelectedOption) return;
-                if (m_CurrentOption != null)
-                {
-                    m_CurrentOption.Reset();
-                }
-                m_CurrentOption = SelectedOption;
-                if (m_CurrentOption != null)
-                {
-                    m_CurrentOption.Initialize();
-                }
-            }
-
-            #region Hotkey Handling
-
-            private readonly Guid m_HotkeyGuid;
-            private string m_Hotkey;
-            private bool m_Registered;
-
-            private void RegisterHotkey()
-            {
-                DynamicHotkeys.RegisterHotkey(m_HotkeyGuid, Hotkey, IncrementSelection);
-                m_Registered = true;
-            }
-
-            private void DeregisterHotkey()
-            {
-                DynamicHotkeys.RemoveHotkey(m_HotkeyGuid);
-                m_Registered = false;
-            }
-
-            private void UpdateHotkey()
-            {
-                if (m_Registered)
-                {
-                    DeregisterHotkey();
-                    RegisterHotkey();
-                }
-            }
-
-            private void IncrementSelection()
-            {
-                if (Options.Count > 0)
-                {
-                    SelectedIndex = (SelectedIndex + 1)%Options.Count;
-                }
-
-                if (SelectedOption != null)
-                {
-                    Player.OsdText.Show(Name + ": " + SelectedOption.Name);
-                }
-            }
-
-            #endregion
-        }
-
-        public class ScriptGroupScript : AudioChainUi<ScriptGroup, ScriptGroupDialog>
+        public class ScriptGroupScript : AudioChainUi<AudioScriptGroup, AudioScriptGroupDialog>
         {
             protected override string ConfigFileName
             {
