+++ Extensions/Framework/Chain/Presets.cs
@@ -16,7 +16,6 @@
 // 
 
 using System;
-using System.Collections.Generic;
 using System.Windows.Forms;
 using YAXLib;
 
@@ -167,29 +166,6 @@ namespace Mpdn.Extensions.Framework.Chain
         {
             var script = new S();
             return new Preset<T, TScript> { Name = (name ?? script.Descriptor.Name), Script = script };
-        }
-    }
-
-    public class PresetCollection< T, TScript> : Chain<T>, INameable
-        where TScript : class, IScript
-    {
-        #region Settings
-
-        public List<Preset<T, TScript>> Options { get; set; }
-
-        [YAXDontSerialize]
-        public string Name { protected get; set; }
-
-        #endregion
-
-        public PresetCollection()
-        {
-            Options = new List<Preset<T, TScript>>();
-        }
-
-        public override T Process(T input)
-        {
-            throw new NotImplementedException();
         }
     }
 }
