+++ Extensions/Framework/RenderChain/ChromaScalerPreset.cs
